\name{attention}
\alias{attention}
\docType{data}
\title{
days_to_death
}
\description{
This is a note to inform package users that the days_to_death variable is also valid for living pateints.
In this case, the value in days_to_death is the number of days since the last follow-up appointment.
}
\format{
  A field included in various data files in the this package.
}
\keyword{datasets}
