% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loadOvarianEsets.R
\name{loadOvarianEsets}
\alias{loadOvarianEsets}
\title{Function to load ovarian cancer expression sets from the Experiment Hub}
\usage{
loadOvarianEsets(remove.duplicates = TRUE, quantile.cutoff = 0,
  rescale = FALSE, min.number.of.genes = 0, min.number.of.events = 0,
  min.sample.size = 0, remove.retracted = TRUE, remove.subsets = TRUE,
  keep.common.only = FALSE, impute.missing = FALSE)
}
\arguments{
\item{remove.duplicates}{remove patients with a Spearman correlation greater than or equal to 0.98 with other patient expression profiles (default TRUE)}

\item{quantile.cutoff}{A nueric between 0 and 1 specifying to remove genes with standard deviation below the required quantile (default 0)}

\item{rescale}{apply centering and scaling to the expression sets (default FALSE)}

\item{min.number.of.genes}{an integer specifying to remove expression sets with less genes than this number (default 0)}

\item{min.number.of.events}{an integer specifying how man survival events must be in the dataset to keep the dataset (default 0)}

\item{min.sample.size}{an integer specifying the minimum number of patients required in an eset (default 0)}

\item{remove.retracted}{remove datasets from retracted papers (default TRUE, currently just PMID17290060 dataset)}

\item{remove.subsets}{remove datasets that are a subset of other datasets (defeault TRUE, currently just PMID19318476)}

\item{keep.common.only}{remove probes not common to all datasets (default FALSE)}

\item{impute.missing}{remove patients from datasets with missing expression values}
}
\value{
a list with 2 elements. The First element named esets contains the datasets. The second element named duplicates contains
a vector with patient IDs for the duplicate patients (those with  Spearman correlation greater than or equal to 0.98 with other patient expression profiles).
}
\description{
This function returns ovarian cancer datasets from the hub and a vector of patients from the datasets that are most likely duplicates
}
\examples{

esetsAndDups = loadOvarianEsets()
}
